requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
tqdm==4.59.0
urllib3==1.26.4
wincertstore==0.2
pycrypto==2.6.1